pages.rs: fix run of orphans packages
This avoid "error: no targets specified (use -h for help)" when `pacman -Qtdq` is empty.

`pacman -Qtdq` can be started without root, and has a return value other than 0 if it finds no packages.
So if it finds packages, it succeeds and goes to the true condition that it deletes packages.
Otherwise do echo. (echo doesn't require root)

Signed-off-by: Peppe289 <gsperanza204@gmail.com>